Fix YAML to use var value instead of var name
Before this change, I got errors on these tasks because it was using the
variable name instead of the variable value. For example, on the "Ensure
language packages are installed" task, I got an error that apt
could not find the `locale_language_packs` package. But it should have
been looking for the _contents_ of the variable, e.g. `language-pack-en`
and `language-pack-en-base`.